MVP Architecture in GroupLoanAccountFragment
