+++ mifosng-android/src/main/AndroidManifest.xml
@@ -88,7 +88,7 @@
 
 
         <activity
-            android:name=".online.GenerateCollectionSheet"
+            android:name=".online.GenerateCollectionSheetActivity"
             android:label="@string/title_activity_generate_collection_sheet"
             android:parentActivityName=".online.DashboardActivity">
             <meta-data
