+++ mifosng-android/src/main/res/layout/activity_generate_collection_sheet.xml
@@ -8,5 +8,5 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.mifos.mifosxdroid.online.GenerateCollectionSheet"
+    tools:context=".online.GenerateCollectionSheetActivity"
     tools:ignore="MergeRootFrame" />
