+++ mifosng-android/src/main/res/layout/fragment_generate_collection_sheet.xml
@@ -26,7 +26,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context="com.mifos.mifosxdroid.online.GenerateCollectionSheet$PlaceholderFragment">
+        tools:context=".online.GenerateCollectionSheetActivity$PlaceholderFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
