+++ mifosng-android/src/main/res/menu/generate_collection_sheet.xml
@@ -6,7 +6,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.mifos.mifosxdroid.online.GenerateCollectionSheet" >
+    tools:context=".online.GenerateCollectionSheetActivity" >
 
     <item android:id="@+id/mItem_search"
         android:title="@string/search"
